requirements.txt
flask>=2.0.0
flask-socketio>=5.0.0
feedparser>=6.0.0
Pillow>=9.0.0
requests>=2.25.0
google-genai>=0.3.0
instagrapi>=2.0.0
python-engineio>=4.0.0
python-socketio>=5.0.0
gunicorn>=20.0.0
eventlet>=0.30.0
